Finish select download
After starting tabel of contents download, navigate to select download fragment.
Here all url on the site with the same host are displayed and their text.
Urls that fit the chapter regex are pre-selected.
User can de-/select chapters that he wants to download.
With the button click the download get started per coroutines(not service)